+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- `useLoadDataEffect()` hook
+
 ## [6.1.1] - 2024-06-13
 
 ### Fixed